requirements.txt
alembic==1.0.11
astroid==2.2.5
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
GeoAlchemy2==0.6.3
gunicorn==19.9.0
imutils==0.5.2
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.0.13
MarkupSafe==1.1.1
marshmallow==2.19.5
mccabe==0.6.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pytesseract==0.2.7
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
